fixed it back. I don't like the helper method because it causes issues in other places. would rather not use it at all.

+++ app/controllers/sessions_controller.rb
@@ -17,7 +17,7 @@ class SessionsController < ApplicationController
   end
 
   def destroy
-    if :current_user
+    if session[:user_id]
       session.delete(:user_id)
       flash[:success] = 'See you!'
     end
